move TimestampFormatter, update geotools
